refactor: #11 폴더 구조 개선 - (anon) 그룹 제거
