+++ app/chatrooms/[id]/components/AdaptChatFromSupabase.ts
