+++ app/chatrooms/[id]/components/ChatInput.tsx
