+++ app/chatrooms/[id]/components/ChatRoomInfo.tsx
