+++ app/chatrooms/[id]/page.module.css
