+++ app/chatrooms/[id]/page.tsx
@@ -7,10 +7,10 @@ import { Message, pMessage } from "./types";
 import styles from "./page.module.css";
 import ChatLog from "./components/ChatLog";
 import ChatInput from "./components/ChatInput";
-import TopTagList from "../../../components/TopTagList";
+import TopTagList from "../../components/TopTagList";
 import ChatSearch from "./components/ChatSearch";
 import ChatRoomInfo from "./components/ChatRoomInfo";
-import { useUserStore } from "@/app/stores/useUserStore";
+import { useUserStore } from "@/stores/useUserStore";
 
 export default function ChatRoom() {
   const params = useParams();
