+++ app/chatrooms/[id]/types.ts
