+++ app/cities/[id]/components/CategoryFilter.tsx
