+++ app/cities/[id]/components/ChatButton.tsx
