+++ app/cities/[id]/components/TourCard.tsx
